Move TestingALM to activitylistmodeltestutils
